quite alot of improvement to the update feature

+++ app/src/main/java/com/jbrightman/thedayto/feature_thedayto/presentation/entry/add_edit_entry/AddEditEntryScreen.kt
@@ -42,7 +42,7 @@ fun AddEditEntryScreen(
     navController: NavController,
     entryColor: Int,
     showBackButton: Boolean,
-    entryDate: Int,
+    entryDate: Long,
     viewModel: AddEditEntryViewModel = hiltViewModel(),
 ) {
     val entryBackgroundAnimatatable = remember {
